+++ remy/api/app.py
@@ -336,6 +336,3 @@ def create_app(
         )
 
     return app
-
-
-app = create_app()